add suspense pages
