+++ package.json
@@ -67,6 +67,7 @@
     "react-geocode": "^0.2.3",
     "react-google-charts": "2",
     "react-i18next": "^11.8.12",
+    "react-lazily": "^0.9.1",
     "react-loading-skeleton": "^2.2.0",
     "react-responsive": "^8.2.0",
     "react-router-dom": "^5.2.0",
@@ -82,8 +83,8 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
     "@testing-library/react-hooks": "^7.0.2",
+    "@testing-library/user-event": "^12.1.10",
     "@types/google-map-react": "^2.1.3",
     "@types/graphql": "^14.5.0",
     "@types/gsap": "^3.0.0",
