+++ src/AppNavigator.tsx
@@ -1,27 +1,29 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useState, lazy, Suspense } from 'react';
+import { lazily } from 'react-lazily';
 import { Switch, Route, useHistory } from 'react-router-dom';
 import { useWeb3React } from '@web3-react/core';
 import ScrollToTop from 'src/hooks/ScrollToTop';
 import usePageTracking from 'src/hooks/usePageTracking';
 import InjectedConnector from 'src/core/connectors/injectedConnector';
 
-import Dashboard from 'src/containers/Dashboard';
-import { StakingEL, StakingELFI } from 'src/containers/Staking';
-import Main from 'src/containers/Main';
-import Governance from 'src/containers/Governance';
+const Dashboard = lazy(() => import('src/containers/Dashboard'))
+const Main = lazy(() => import('src/containers/Main'))
+const Governance = lazy(() => import('src/containers/Governance'))
+const { StakingEL, StakingELFI } = lazily(() => import('src/containers/Staking'));
+const LPStaking = lazy(() => import('src/containers/LPStaking'));
+const RewardPlan = lazy(() => import('src/containers/RewardPlan'));
+const MarketDetail = lazy(() => import('src/containers/MarketDetails'));
+const PortfolioDetail = lazy(() => import('src/containers/PortfolioDetail'));
 
 import 'src/stylesheet/public.scss';
 import 'src/stylesheet/pc.scss';
-// import 'src/stylesheet/tablet.scss';
 import 'src/stylesheet/mobile.scss';
-import Navigation from 'src/components/Navigation';
-import Footer from 'src/components/Footer';
+
+const Navigation = lazy(() => import('src/components/Navigation'))
+const Footer = lazy(() => import('src/components/Footer'))
+
 import getLocalLanauge from 'src/utiles/getLocalLanguage';
 import LanguageProvider from 'src/providers/LanguageProvider';
-import LPStaking from 'src/containers/LPStaking';
-import RewardPlan from 'src/containers/RewardPlan';
-import MarketDetail from 'src/containers/MarketDetails';
-import PortfolioDetail from 'src/containers/PortfolioDetail';
 import useMediaQueryType from 'src/hooks/useMediaQueryType';
 import MediaQuery from 'src/enums/MediaQuery';
 import { isMetamask, isWalletConnector } from './utiles/connectWallet';
@@ -81,29 +83,48 @@ const AppNavigator: React.FC = () => {
       <Switch>
         <Route path="/:lng">
           <LanguageProvider>
-            <Navigation
-              hamburgerBar={hamburgerBar}
-              setHamburgerBar={setHamburgerBar}
-            />
-            <Route exact path="/:lng/staking/LP" component={LPStaking} />
-            <Route exact path="/:lng/staking/EL" component={StakingEL} />
-            <Route exact path="/:lng/staking/ELFI" component={StakingELFI} />
-            <Route exact path="/:lng/deposit" component={Dashboard} />
-
-            <Route exact path="/:lng/governance" component={Governance} />
-            <Route
-              exact
-              path="/:lng/portfolio/:id"
-              component={PortfolioDetail}
-            />
-            <Route
-              exact
-              path="/:lng/rewardplan/:stakingType"
-              component={RewardPlan}
-            />
-            <Route exact path="/:lng/deposits/:id" component={MarketDetail} />
-            <Route exact path="/:lng" component={Main} />
-            <Footer />
+              <Suspense fallback={null}>
+                <Navigation
+                  hamburgerBar={hamburgerBar}
+                  setHamburgerBar={setHamburgerBar}
+                />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Route exact path="/:lng/staking/LP" component={LPStaking} />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Route exact path="/:lng/staking/EL" component={StakingEL} />
+                <Route exact path="/:lng/staking/ELFI" component={StakingELFI} />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Route exact path="/:lng/governance" component={Governance} />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Route
+                  exact
+                  path="/:lng/portfolio/:id"
+                  component={PortfolioDetail}
+                />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Route
+                  exact
+                  path="/:lng/rewardplan/:stakingType"
+                  component={RewardPlan}
+                />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Route exact path="/:lng/deposit/:id" component={MarketDetail} />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Route exact path="/:lng/deposit" component={Dashboard} />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Route exact path="/:lng" component={Main} />
+              </Suspense>
+              <Suspense fallback={null}>
+                <Footer />
+              </Suspense>
           </LanguageProvider>
         </Route>
         <Route path="/" component={LanguageDetctionPage} />
