+++ src/components/Staking/StakingProgress.tsx
