+++ src/components/TvlCounter.tsx
@@ -1,10 +1,10 @@
 import CountUp from 'react-countup';
-import useMediaQueryType from 'src/hooks/useMediaQueryType';
-import HeaderCircle from 'src/assets/images/title-circle.png';
+import HeaderCircle from 'src/assets/images/deposit/title-circle.svg';
 import useTvl from 'src/hooks/useTvl';
 import Skeleton from 'react-loading-skeleton';
 import { useMemo } from 'react';
 import MediaQuery from 'src/enums/MediaQuery';
+import useMediaQueryType from 'src/hooks/useMediaQueryType';
 
 const usdFormatter = new Intl.NumberFormat('en', {
   style: 'currency',
