+++ src/utiles/lazyImage.tsx
@@ -1,10 +1,13 @@
+import { CSSProperties } from "react";
+
 export interface IProps {
   src: string,
-  name: string
+  name: string,
+  style?: CSSProperties
 }
 
-const LazyImage:React.FC <IProps> = ({src, name}) => {
-  return <img src={src} alt={name} />
+const LazyImage:React.FC <IProps> = ({src, name, style}) => {
+  return <img src={src} className={name} alt={name} style={style} />
 }
 
 export default LazyImage;
